Mis pedidos

+++ ecommerce-api/src/main/java/com/jbrod/ecommerce_api/repositorios/pedido/EstadoPedidoRepository.java
@@ -5,6 +5,10 @@ import com.jbrod.ecommerce_api.modelos.pedidos.EstadoPedido;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface EstadoPedidoRepository extends JpaRepository<EstadoPedido, Long> {
+public interface EstadoPedidoRepository extends JpaRepository<EstadoPedido, Integer> {
+
+
 }
